A very lightweight DOM2 implementation. Only the most essential classes have been implemented. These should, however, be sufficient for 90% of all XML applications. Most of the node classes were copied from Helge's Skyrix XML framework; with kind permission.
Note that the classes simply implement the methods described in DOM2-Core. I have passed on creating a formal DOM protocol set that concrete classes implement. If you are using stuff from Skyrix XML you can simply cast the EDXML classes to the corresponding DOM protocols defined there...

Missing for a release are: (a) a class that can write DOM objects to an XML file and convenience methods in XMLNode to use these, and (b) documentation.